Extract syscall_lseek() handler
